PD-2865 Change Int32 to Long in DatasetUploadReport; Switch from NET45 to NET48

+++ Knoema.Client/Upload/DimensionChanges.cs
@@ -5,9 +5,9 @@ namespace Knoema.Upload
 	public class DimensionChanges
 	{
 		public List<string> AddedFields { get; set; }
-		public int AddedMembers { get; set; }
-		public int UpdatedMembers { get; set; }
-		public int TotalMembersInUpdate { get; set; }
+		public long AddedMembers { get; set; }
+		public long UpdatedMembers { get; set; }
+		public long TotalMembersInUpdate { get; set; }
 
 		public DimensionChanges()
 		{
